requirements.txt
requests==2.28.1
python-dotenv==1.0.0
PyGithub==1.56.0
